requirements.txt
flask
twilio
python-dotenv
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
Pillow
sqlalchemy
psycopg2-binary
google-cloud-bigquery
